Make liba2ps an internal library
This was motivated by wanting to move some sources into a different
library (automake’s recommended workaround for compiling with different
CFLAGS), to avoid compiler warnings for sources generated by flex and yacc,
as well as pristine GNU regex.

Hence, remove the sample project in contrib/sample, and the version number
for liba2ps.

libtool is still required to build gnulib, and we also use it for liba2ps to
avoid needing to repeat libraries on the link line as we would with purely
static libraries.