Use gnulib emit_bug_reporting_address() (thanks, Benno Schulenberg)
This uses the gnulib message catalog, and fixes a problem where gettext does
not understand C string concatenation.

As a result, remove the unused ability to output --help output to a stream
other than stdout, as emit_bug_reporting_address can only output to stdout.

Change a couple of other calls that pass stdout as a hard-wired parameter to
output functions to use the equivalent stdout-only APIs instead.