Use libpaper’s paper sizes
Get paper sizes from libpaper, instead of having a built-in list in
a2ps.cfg. Keep just one entry, Quarto, which is not defined in libpaper, as
an example, and for backwards compatibility.

Test this by removing the definition of A4 from tests/a2ps-tst.cfg.in.

Also fix parsing of Medium configuration lines in confg.gperf to require
unsigned dimensions.